requirements.txt
Flask==3.0.0
blueprint==3.4.2
SQLAlchemy==2.0.22
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
PyJWT-2.8.0